--- genshin-map-ability/genshin-map-ability-auth/src/main/java/site/yuanshen/auth/AuthServerApplication.java
@@ -1,17 +1,0 @@
-package site.yuanshen.auth;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.context.annotation.ComponentScan;
-
-@EnableDiscoveryClient
-@SpringBootApplication
-@ComponentScan(basePackages = {"site.yuanshen.auth", "site.yuanshen.common.core.utils"})
-public class AuthServerApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AuthServerApplication.class, args);
-    }
-
-}
