--- genshin-map-ability/genshin-map-ability-auth/src/main/java/site/yuanshen/auth/config/JwtAccessToken.java
@@ -1,21 +1,0 @@
-package site.yuanshen.auth.config;
-
-import org.springframework.security.jwt.JwtHelper;
-import org.springframework.security.oauth2.common.*;
-import org.springframework.security.oauth2.provider.OAuth2Authentication;
-import org.springframework.security.oauth2.provider.token.store.JwtAccessTokenConverter;
-import org.springframework.stereotype.Component;
-
-import java.util.Date;
-import java.util.LinkedHashMap;
-import java.util.Map;
-
-@Component
-public class JwtAccessToken extends JwtAccessTokenConverter {
-    @Override
-    public OAuth2AccessToken enhance(OAuth2AccessToken accessToken, OAuth2Authentication authentication) {
-        DefaultOAuth2AccessToken defaultOAuth2AccessToken = new DefaultOAuth2AccessToken(accessToken);
-        defaultOAuth2AccessToken.setExpiration(new Date(System.currentTimeMillis() + 1000 * 1800 ));
-        return super.enhance(defaultOAuth2AccessToken, authentication);
-    }
-}
