--- genshin-map-ability/genshin-map-ability-auth/src/main/java/site/yuanshen/auth/model/dto/RoleSecurityDto.java
@@ -1,41 +1,0 @@
-package site.yuanshen.auth.model.dto;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import site.yuanshen.common.core.utils.BeanUtils;
-import site.yuanshen.data.enums.RoleEnum;
-
-/**
- * 角色Dto
- *
- * @author Moment
- */
-@Data
-@NoArgsConstructor
-public class RoleSecurityDto implements GrantedAuthority {
-
-    /**
-     * 角色名
-     */
-    private String name;
-
-    /**
-     * 角色代码（英文大写）
-     */
-    private String code;
-
-    /**
-     * 角色层级（越大级别越高）
-     */
-    private Integer sort;
-
-    @Override
-    public String getAuthority() {
-        return code;
-    }
-
-    public RoleSecurityDto(String code) {
-        BeanUtils.copyProperties(RoleEnum.valueOf(code), this);
-    }
-}
