--- genshin-map-ability/genshin-map-ability-auth/src/main/resources/bootstrap.yml
@@ -1,17 +1,0 @@
-spring:
-  application:
-    name: oauth2-auth
-  profiles:
-    active: "@runtime-profile@"
-
-server:
-  port: 8100
-
-#logging:
-#  level:
-#    org:
-#      springframework:
-#        security: DEBUG
-
-
-
