+++ genshin-map-api/genshin-map-api-core/genshin-map-api-core-core/pom.xml
@@ -26,6 +26,29 @@
             <artifactId>genshin-map-data-system-mapper</artifactId>
             <version>${project.version}</version>
         </dependency>
+
+        <!--START Auth用依赖-->
+        <dependency>
+            <groupId>com.nimbusds</groupId>
+            <artifactId>nimbus-jose-jwt</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-web</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.springframework.cloud</groupId>
+            <artifactId>spring-cloud-starter-oauth2</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.springframework.cloud</groupId>
+            <artifactId>spring-cloud-starter-security</artifactId>
+        </dependency>
+        <!--END Auth用依赖-->
+
     </dependencies>
 
 </project>
