+++ genshin-map-api/genshin-map-api-core/genshin-map-api-core-core/src/main/resources/oauth2.jks
