--- genshin-map-api/genshin-map-api-system/genshin-map-api-system-core/pom.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>genshin-map-api-system</artifactId>
-        <groupId>site.yuanshen</groupId>
-        <version>${gsapi.version}</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>genshin-map-api-system-core</artifactId>
-
-    <dependencies>
-        <dependency>
-            <groupId>site.yuanshen</groupId>
-            <artifactId>genshin-map-commons-core</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>site.yuanshen</groupId>
-            <artifactId>genshin-map-data-system-mapper</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-    </dependencies>
-</project>
