--- genshin-map-api/genshin-map-api-system/genshin-map-api-system-core/src/main/java/site/yuanshen/api/system/ApiSystemApplication.java
@@ -1,15 +1,0 @@
-package site.yuanshen.api.system;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-@EnableDiscoveryClient
-@SpringBootApplication
-public class ApiSystemApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ApiSystemApplication.class, args);
-    }
-
-}
