--- genshin-map-api/genshin-map-api-system/genshin-map-api-system-core/src/main/resources/banner.txt
@@ -1,22 +1,0 @@
-###############################################WEEEEKj EEEEEK#########
-######WGGDDDDDGGW################WWGGDDDDDGGW##GKEEEK. EEEEEEW########
-##WGEKEEEEK.EEEEEEEEEEEEEEEEEKEEEEEEEEEEEEEEEEEEEEEEt   EEEEEEDW######
-##WDKEEEEEL .EEEEEEEEEEEEEEEEEEEEEEEEK. KEEEEEEKEEEE    ,EEEEEEKEDWW##
-##GEKEK                                    .EEEEE.         LEEEEEEDEGW
-##WEEEEEE.   .K;fEKEKD   GKKf,DEEEEEEKi.E  GEj.tEEEKK  jKEKL iEKEEEKEG
-##WEEEEEE.   .K                 E.          K:                  DEEEKG
-###EKEEEE.   .K    DEEEEEEK:   .KD  :;      .;    KEG   KEi    DEEEEED
-###EEEEEE.   .K           t:   .KE  .;  .EtjE:    ,       ,    EEEEEEG
-###DEEEEE.   .K    EKKKEKKW:   .KE  .;  .t  t,    KE     Ki    EEEEEEG
-##WKEEEEE    GE   :;;.   .;    :KE  .;  .t  i,    EEK. DEEt    EEEEKKW
-#WGKEEEEE    EE  LKEK;   :K,..jEEE  .;  .t  D:                 DEEEEKW
-#WDKEEEEE    EKEj. EK;   :Ki .tKKD  :;  .j  K,   jKEK. DKEE    DEEEEKW
-WGKEEEEK.  :L        j   i          K,  .KEEK;.DEKKD    .KKKK,.DEEEEDW
-WDDEEEK  :KK:  LEEEEK;   :EEEKEi   KK,   KEEEEEEEt        :DEEEEEEEEG#
-WDDEK.EKEEEEEEEEE:;EK,   EEE:iEEEEEEK:  DEEEKEEEEEEE    :EEEEKEEDEDGW#
-WGDDEEEEEEEEEEEEEEEEK..EEEEEEEEEEEEEK. EEEEEEEEEEEEE,   EEEEEEEDWW####
-#WWGDDEEEEEEKKKKEEEEEEEEEEEEKDKDEKEKEEEEEEKEDGGKEEEEE  :EEEEEDW#######
-##############WWGDEDKEEEDKEDGGWWGDDEKKDKEEDDW##EEEEEK  DEEEEKW########
-###############################################WDEEEKi EEEEEK#########
-###############################################WKEEEEE.EEEEDD#########
-################################################DKDEED:KEEDEG#########
