--- genshin-map-api/genshin-map-api-system/genshin-map-api-system-core/src/main/resources/bootstrap.yml
@@ -1,7 +1,0 @@
-spring:
-  application:
-    name: api-system
-  profiles:
-    active: "@runtime-profile@"
-server:
-  port: 8091
