--- genshin-map-api/genshin-map-api-system/pom.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>genshin-map-api</artifactId>
-        <groupId>site.yuanshen</groupId>
-        <version>${gsapi.version}</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>genshin-map-api-system</artifactId>
-    <packaging>pom</packaging>
-    <modules>
-        <module>genshin-map-api-system-core</module>
-    </modules>
-
-    <dependencies>
-
-    </dependencies>
-
-</project>
