+++ genshin-map-data/genshin-map-data-system/genshin-map-data-system-model/src/main/java/site/yuanshen/data/dto/SysUserSecurityDto.java
@@ -1,8 +1,11 @@
 package site.yuanshen.data.dto;
 
 import lombok.Data;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
 import site.yuanshen.data.entity.SysUser;
 
+import java.util.Collection;
 import java.util.List;
 
 /**
@@ -11,7 +14,7 @@ import java.util.List;
  * @author Moment
  */
 @Data
-public class SysUserSecurityDto {
+public class SysUserSecurityDto  implements UserDetails {
 
     /**
      * 用户ID
@@ -47,5 +50,37 @@ public class SysUserSecurityDto {
      * 角色
      */
     private List<SysRoleDto> roleDtoList;
+
+
+
+    @Override
+    public String getUsername() {
+        return username;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return roleDtoList;
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
 
 }
